Use BigInt to send values

+++ src/NiMPC/Communication.nim
@@ -1,4 +1,6 @@
+import bigints
+
 type Party* = ref object
 
-method send*(sender: Party, recipient: Party, value: int) =
+method send*(sender: Party, recipient: Party, value: BigInt) =
   discard
